audispot_content_writer/requirements.txt
openai
openai-agents
youtube_transcript_api
python-dotenv
streamlit 